Fixed compiler warnings
client/readline.cc:
  Initialize not used variable (to kill wrong compiler warning)
mysql-test/suite/handler/aria.result:
  Updated test result
mysql-test/suite/handler/handler.inc:
  Changed index to ensure rows are in a fixed order
mysql-test/suite/handler/heap.result:
  Updated test result
mysql-test/suite/handler/innodb.result:
  Updated test result
mysql-test/suite/handler/myisam.result:
  Updated test result
plugin/handler_socket/handlersocket/Makefile.am:
  Use CXX flags to compile
sql/filesort.cc:
  Initialize variable that may be used
sql/log.cc:
  Initialize not used variable (to kill wrong compiler warning)
sql/opt_range_mrr.cc:
  Fixed cast to avoid compiler warning
storage/xtradb/fil/fil0fil.c:
  Added cast to avoid compiler warning

+++ storage/xtradb/fil/fil0fil.c
@@ -3313,7 +3313,7 @@ skip_info:
 				}
 
 				if (page_is_corrupt) {
-					fprintf(stderr, " [errp:%lld]", offset / UNIV_PAGE_SIZE);
+                                  fprintf(stderr, " [errp:%lld]", (longlong) (offset / UNIV_PAGE_SIZE));
 
 					/* cannot treat corrupt page */
 					goto skip_write;
